+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.25.0] - 2024-03-03
 
 ### Added
 - A shortcut in the preferences for Android 13 app language settings
@@ -460,7 +460,8 @@ icon on the main screen
 - "Library is not accessible" during the connection, when the entered root URL doesn't match the 
 `PHOTOPRISM_SITE_URL` server config value 
 
-[Unreleased]: https://github.com/Radiokot/photoprism-android-client/compare/1.24.1(38)...HEAD
+[Unreleased]: https://github.com/Radiokot/photoprism-android-client/compare/1.25.0(39)...HEAD
+[1.25.0]: https://github.com/Radiokot/photoprism-android-client/compare/1.24.1(38)...1.25.0(39)
 [1.24.1]: https://github.com/Radiokot/photoprism-android-client/compare/1.24.0(37)...1.24.1(38)
 [1.24.0]: https://github.com/Radiokot/photoprism-android-client/compare/1.23.0(36)...1.24.0(37)
 [1.23.0]: https://github.com/Radiokot/photoprism-android-client/compare/1.22.0(35)...1.23.0(36)